requirements.txt
# LIST ALL PYTHON PACKAGES NEEDED TO RUN THIS PROGRAM IN BELOW FORMAT WITHOUT #

#package 1
#Package 2

pytest
python-dotenv
sendgrid==5.6.0
gspread 
oauth2client
pprint